refactor(map): 提取内部包并整理静态变量结构
