+++ agent/go-service/map-tracker/const.go
@@ -6,22 +6,6 @@ const (
 	WORK_H = 720
 )
 
-// Location inference configuration
-const (
-	// Mini-map crop area
-	LOC_CENTER_X = 108
-	LOC_CENTER_Y = 111
-	LOC_RADIUS   = 40
-)
-
-// Rotation inference configuration
-const (
-	// Pointer crop area
-	ROT_CENTER_X = 108
-	ROT_CENTER_Y = 111
-	ROT_RADIUS   = 12
-)
-
 // Big map zoom button configuration
 const (
 	ZOOM_BUTTON_AREA_X    = 0.95 * WORK_W
@@ -48,14 +32,6 @@ const (
 	BIG_MAP_PICK_RETRY = 10
 )
 
-// Time-series empirical optimization configuration
-const (
-	PENDING_TAKEOVER_TIME_MS         = 1000
-	PENDING_TAKEOVER_COUNT_THRESHOLD = 3
-	CONVINCED_DISTANCE_THRESHOLD     = 30
-	CONVINCED_VALID_TIME_MS          = 2000
-)
-
 // Resource paths
 const (
 	MAP_BBOX_DATA_PATH     = "data/MapTracker/map_bbox_data.json"
@@ -73,44 +49,8 @@ const (
 
 // Move action fine approach configuration
 const (
-	FINE_APPROACH_FINAL_TARGET       = "FinalTarget"
-	FINE_APPROACH_ALL_TARGETS        = "AllTargets"
-	FINE_APPROACH_NEVER              = "Never"
 	FINE_APPROACH_COMPLETE_THRESHOLD = 0.5
 )
-
-// MapTrackerInfer parameters default values
-var DEFAULT_INFERENCE_PARAM = MapTrackerInferParam{
-	MapNameRegex: "^map\\d+_lv\\d+$",
-	Precision:    0.5,
-	Threshold:    0.4,
-}
-
-// MapTrackerInfer parameters for MapTrackerMove action default values
-// (MapNameRegex is omitted here since MapTrackerMove always sets it)
-var DEFAULT_INFERENCE_PARAM_FOR_MOVE = MapTrackerInferParam{
-	Precision: 0.7,
-	Threshold: 0.3,
-}
-
-// MapTrackerBigMapInfer parameters default values
-var DEFAULT_BIG_MAP_INFERENCE_PARAM = MapTrackerBigMapInferParam{
-	MapNameRegex: "^map\\d+_lv\\d+$",
-	Threshold:    0.5,
-}
-
-// MapTrackerMove parameters default values
-var DEFAULT_MOVING_PARAM = MapTrackerMoveParam{
-	FineApproach:           FINE_APPROACH_FINAL_TARGET,
-	MapNameMatchRule:       "^%s(_tier_\\w+)?$",
-	ArrivalThreshold:       2.5,
-	ArrivalTimeout:         60000,
-	RotationLowerThreshold: 7.5,
-	RotationUpperThreshold: 60.0,
-	SprintThreshold:        20.0,
-	StuckThreshold:         2000,
-	StuckTimeout:           10000,
-}
 
 // Win32 action related codes
 const (
